Organize application into subpackages
