+++ watermark_remover/__init__.py
@@ -1,0 +1,9 @@
+"""Top-level package for the Watermark Remover application."""
+
+__all__ = [
+    "gui",
+    "download",
+    "threads",
+    "inference",
+    "utils",
+]
