+++ watermark_remover/gui/dialogs/batch_grid_dialog.py
