+++ watermark_remover/gui/dialogs/suggestion_dialog.py
